FUSE dummy parameters

This dataset contains an example of FUSE parameter set. It consists of a data.frame with 24 columns (1 each sampled parameter value).

Format

A data frame with 1 observations on the following 24 variables.

rferr_addadditive rainfall error (mm)

rferr_mltmultiplicative rainfall error (-)

maxwatr_1depth of the upper soil layer (mm)

maxwatr_2depth of the lower soil layer (mm)

fractenfraction total storage in tension storage (-)

frchznefraction tension storage in recharge zone (-)

fprimqbfraction storage in 1st baseflow reservoir (-)

rtfrac1fraction of roots in the upper layer (-)

percrtepercolation rate (mm day-1)

percexppercolation exponent (-)

sacpmltSACRAMENTO model percolation multiplier for dry soil layer (-)

sacpexpSACRAMENTO model percolation exponential coefficient for dry soil layer (-)

percfracfraction of percolation to tension storage (-)

iflwrteinterflow rate (mm day-1)

basertebaseflow rate (mm day-1)

qb_powrbaseflow exponent (-)

qb_prmsbaseflow depletion rate (day-1)

qbrate_2abaseflow depletion rate 1st reservoir (day-1)

qbrate_2bbaseflow depletion rate 2nd reservoir (day-1)

sareamaxmaximum saturated area (-)

axv_bexpARNO/VIC model "b" exponent (-)

loglambmean value of the topographic index (m)

tishapeshape parameter for the topographic index Gamma distribution (-)

timedelaytime delay in runoff (days)

Details

Data set: fuse_parameters

Examples

data(fuse_parameters)
str(fuse_parameters)

#> 'data.frame':	1 obs. of  24 variables:
#>  $ rferr_add: num 0
#>  $ rferr_mlt: num 1
#>  $ maxwatr_1: num 162
#>  $ maxwatr_2: num 1404
#>  $ fracten  : num 0.194
#>  $ frchzne  : num 0.235
#>  $ fprimqb  : num 0.324
#>  $ rtfrac1  : num 0.274
#>  $ percrte  : num 440
#>  $ percexp  : num 18.7
#>  $ sacpmlt  : num 145
#>  $ sacpexp  : num 4.19
#>  $ percfrac : num 0.33
#>  $ iflwrte  : num 87.4
#>  $ baserte  : num 242
#>  $ qb_powr  : num 7.48
#>  $ qb_prms  : num 0.0142
#>  $ qbrate_2a: num 0.16
#>  $ qbrate_2b: num 0.174
#>  $ sareamax : num 0.597
#>  $ axv_bexp : num 2.92
#>  $ loglamb  : num 6.84
#>  $ tishape  : num 4.97
#>  $ timedelay: num 4.19