This function generates parameter sets for FUSE models (see ranges suggested by Clark et al., 2011).

This function generates parameter sets for FUSE models (see ranges suggested by Clark et al., 2011).

generateParameters(NumberOfRuns, SamplingType = "LHS", rferr_add = NULL,
  rferr_mlt = NULL, frchzne = NULL, fracten = NULL, maxwatr_1 = NULL,
  percfrac = NULL, fprimqb = NULL, qbrate_2a = NULL, qbrate_2b = NULL,
  qb_prms = NULL, maxwatr_2 = NULL, baserte = NULL, rtfrac1 = NULL,
  percrte = NULL, percexp = NULL, sacpmlt = NULL, sacpexp = NULL,
  iflwrte = NULL, axv_bexp = NULL, sareamax = NULL, loglamb = NULL,
  tishape = NULL, qb_powr = NULL, timedelay = NULL,
  params2remove = NULL)

Arguments

NumberOfRuns

number of samples to generate, can be any integer

SamplingType

sampling procedure to use, can be "URS" or "LHS"

rferr_add

range of the additive rainfall error (mm), default is c(0,0)

rferr_mlt

range of the multiplicative rainfall error (-), default is c(1,1)

frchzne

range of the fraction tension storage in recharge zone (-), default is c(0.05,0.95)#'

fracten

range of the fraction total storage in tension storage (-), default is c(0.05,0.95)

maxwatr_1

range of the depth of the upper soil layer (mm), default is c(25,500)

percfrac

range of the fraction of percltn to tension storage (-), default is c(0.05,0.95)

fprimqb

range of the fraction storage in 1st baseflow reservoir (-), default is c(0.05,0.95)

qbrate_2a

range of the baseflow depletion rate 1st reservoir (day-1), default is c(0.001,0.25)

qbrate_2b

range of the baseflow depletion rate 2nd reservoir (day-1), default is c(0.001,0.25)

qb_prms

range of the baseflow depletion rate (day-1), default is c(0.001,0.25)

maxwatr_2

range of the depth of the lower soil layer (mm), default is c(50,5000)

baserte

range of the baseflow rate (mm day-1), default is c(0.001,1000)

rtfrac1

range of the fraction of roots in the upper layer (-), default is c(0.05,0.95)

percrte

range of the percolation rate (mm day-1), default is c(0.01,1000)

percexp

range of the percolation exponent (-), default is c(1,20)

sacpmlt

range of the SAC model percltn mult for dry soil layer (-), default is c(1,250)

sacpexp

range of the SAC model percltn exp for dry soil layer (-), default is c(1,5)

iflwrte

range of the interflow rate (mm day-1), default is c(0.01,1000)

axv_bexp

range of the ARNO/VIC b exponent (-), default is c(0.001,3)

sareamax

range of the maximum saturated area (-), default is c(0.05,0.95)

loglamb

range of the mean value of the topographic index (m), default is c(5,10)

tishape

range of the shape param for the topo index Gamma dist (-), default is c(2,5)

qb_powr

range of the baseflow exponent (-), default is c(1,10)

timedelay

range of the time delay in runoff (days), default is c(0.01,5)

params2remove

vector with names of parameters to remove from the latin hypercube.

Examples

## Not run: ------------------------------------
# # For reproducible results, use set.seed() before running this function.
# set.seed(123)
# parameters <- generateParameters(1)
## ---------------------------------------------