FUSE model structures (numeric)

This object contains the list of FUSE model structures implemented by Clark et al. (2011). This is available as numerical codes (modlist) or strings (modliststring). The variables are as follows:

mid = model id number

rferr = type of rainfall error

arch1 = architecture of the upper soil layer

arch2 = architecture of the lower soil layer

qsurf = surface runoff

qperc = vertical drainage

esoil = evapotranspiration

qintf = interflow

q_tdh = routing

Each row identifies a model structure and each column identifies a different element of the model structure. For each element of the model structure, several model decisions can be made (see Clark et al. 2011 for details):

Rainfall error (rferr)

additive_e (additive rainfall error) = 11

multiplc_e (multiplicative rainfall error) = 12

Architecture of the upper soil layer (arch1)

onestate_1 (Single state: upper layer defined by a single state variable) = 21

tension1_1 (Separate tension storage: upper layer broken up into tension and free storage) = 22

tension2_1 (Cascading buckets: tension storage sub-divided into recharge and excess) = 23

Architecture of the lower soil layer (arch2)

fixedsiz_2 (baseflow reservoir of fixed size) = 31

tens2pll_2 (tension reservoir plus two parallel tanks) = 32

unlimfrc_2 (baseflow reservoir of unlimited size (0-HUGE), frac rate) = 33

unlimpow_2 (baseflow reservoir of unlimited size (0-HUGE), power recession) = 34

Runoff (qsurf)

arno_x_vic (Unsaturated zone Pareto: ARNO/Xzang/VIC parameterization (upper zone control)) = 41

prms_varnt (Unsaturated zone linear: PRMS variant (fraction of upper tension storage)) = 42

tmdl_param (Saturated zone topographic: TOPMODEL parameterization (only valid for TOPMODEL qb)) = 43

Percolation (qperc)

perc_f2sat (Drainage above field capacity: water from (field cap to sat) avail for percolation) = 51

perc_lower (Gravity drainage: percolation defined by moisture content in lower layer (SAC)) = 52

perc_w2sat (Saturated zone control: water from wilt point to saturation is available for percolation) = 53

Evaporation (esoil)

rootweight (root weighting) = 61

sequential (sequential evaporation model) = 62

Interflows (qintf)

intflwnone (interflow denied) = 71

intflwsome (interflow allowed) = 72

Routing (q_tdh)

no_routing (routing denied) = 81

rout_gamma (routing allowed using Gamma distribution with shape parameter = 2.5) = 82

The function readmd() returns the model components corresponding to a given id. A similar function (readmd2var()) reads the components as either character or numeric vector.

On Clark's model list, the parent models (Topmodel, VIC, PRMS and Sacramento) have id = 60, 230, 342 and 426 respectively.

For example, model 5 is identified by the following numeric codes: data(modlist) modlist[5,]

Format

A data frame with 1248 rows and 9 variables

Details

Data set: modlist