This dataset consists of a data.frame with 24 columns (1 each sampled parameter value)

data(parameters)

Format

A data frame with 1 observations on the following 24 variables.

rferr_add

additive rainfall error (mm)

rferr_mlt

multiplicative rainfall error (-)

maxwatr_1

depth of the upper soil layer (mm)

maxwatr_2

depth of the lower soil layer (mm)

fracten

fraction total storage in tension storage (-)

frchzne

fraction tension storage in recharge zone (-)

fprimqb

fraction storage in 1st baseflow reservoir (-)

rtfrac1

fraction of roots in the upper layer (-)

percrte

percolation rate (mm day-1)

percexp

percolation exponent (-)

sacpmlt

SACRAMENTO model percolation multiplier for dry soil layer (-)

sacpexp

SACRAMENTO model percolation exponential coefficient for dry soil layer (-)

percfrac

fraction of percolation to tension storage (-)

iflwrte

interflow rate (mm day-1)

baserte

baseflow rate (mm day-1)

qb_powr

baseflow exponent (-)

qb_prms

baseflow depletion rate (day-1)

qbrate_2a

baseflow depletion rate 1st reservoir (day-1)

qbrate_2b

baseflow depletion rate 2nd reservoir (day-1)

sareamax

maximum saturated area (-)

axv_bexp

ARNO/VIC model "b" exponent (-)

loglamb

mean value of the topographic index (m)

tishape

shape parameter for the topographic index Gamma distribution (-)

timedelay

time delay in runoff (days)

Details

This is an example dataset.

Examples

data(parameters)
str(parameters)

#> 'data.frame':	1 obs. of  24 variables:
#>  $ rferr_add: num 0
#>  $ rferr_mlt: num 1
#>  $ maxwatr_1: num 162
#>  $ maxwatr_2: num 1404
#>  $ fracten  : num 0.194
#>  $ frchzne  : num 0.235
#>  $ fprimqb  : num 0.324
#>  $ rtfrac1  : num 0.274
#>  $ percrte  : num 440
#>  $ percexp  : num 18.7
#>  $ sacpmlt  : num 145
#>  $ sacpexp  : num 4.19
#>  $ percfrac : num 0.33
#>  $ iflwrte  : num 87.4
#>  $ baserte  : num 242
#>  $ qb_powr  : num 7.48
#>  $ qb_prms  : num 0.0142
#>  $ qbrate_2a: num 0.16
#>  $ qbrate_2b: num 0.174
#>  $ sareamax : num 0.597
#>  $ axv_bexp : num 2.92
#>  $ loglamb  : num 6.84
#>  $ tishape  : num 4.97
#>  $ timedelay: num 4.19